tests/wpt/tests/IndexedDB/idbtransaction-objectStore-finished.html

<!doctype html>
<meta charset=utf-8>
<title>IndexedDB: IDBTransaction objectStore() when transaction is finished</title>
<link rel="help" href="https://w3c.github.io/IndexedDB/#dom-idbtransaction-objectstore">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/support.js"></script>
<script>

indexeddb_test(
  (t, db) => {
    db.createObjectStore('store');
  },
  (t, db) => {
    const tx = db.transaction('store', 'readonly', {durability: 'relaxed'});
    tx.abort();
    assert_throws_dom('InvalidStateError', () => tx.objectStore('store'),
                      'objectStore() should throw if transaction is finished');
    t.done();
  },
  'IDBTransaction objectStore() behavior when transaction is finished'
);

</script>
